requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-RESTX==1.1.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7